# laravel-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API – Aplikacija za izložbe fotografija",
        "description": "Swagger dokumentacija Laravel API-ja",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lokalno (Docker)"
        }
    ],
    "paths": {
        "/api/registracija": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Registracija korisnika (posetilac)",
                "operationId": "01f458cfcfa16b238b86777cb1c62aea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ime",
                                    "prezime",
                                    "email",
                                    "lozinka",
                                    "lozinka_confirmation"
                                ],
                                "properties": {
                                    "ime": {
                                        "type": "string",
                                        "example": "Pera"
                                    },
                                    "prezime": {
                                        "type": "string",
                                        "example": "Perić"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "pera@test.com"
                                    },
                                    "lozinka": {
                                        "type": "string",
                                        "example": "pera123"
                                    },
                                    "lozinka_confirmation": {
                                        "type": "string",
                                        "example": "pera123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Korisnik registrovan"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            }
        },
        "/api/prijava": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Prijava korisnika",
                "operationId": "534ba388340e751c9ec73ede1668b3dd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "lozinka"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "pera@test.com"
                                    },
                                    "lozinka": {
                                        "type": "string",
                                        "example": "pera123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ulogovan"
                    },
                    "401": {
                        "description": "Pogrešan email ili lozinka"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            }
        },
        "/api/odjava": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Odjava korisnika",
                "operationId": "21de471e1d40554ea5e664c48ce4b626",
                "responses": {
                    "200": {
                        "description": "Odjavljen"
                    },
                    "401": {
                        "description": "Niste autentifikovani"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/korisnik": {
            "get": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Vraća ulogovanog korisnika",
                "operationId": "baaa5dbf2bcac8cd324495d1aad5d660",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Niste autentifikovani"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/zaboravljena-lozinka": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Reset lozinke (direktna promena)",
                "operationId": "e64ae69e3d1bfc9a55daccbf463db2c0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "nova_lozinka",
                                    "nova_lozinka_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "pera@test.com"
                                    },
                                    "nova_lozinka": {
                                        "type": "string",
                                        "example": "nova12345"
                                    },
                                    "nova_lozinka_confirmation": {
                                        "type": "string",
                                        "example": "nova12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lozinka promenjena"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            }
        },
        "/api/fotografije": {
            "get": {
                "tags": [
                    "Fotografije"
                ],
                "summary": "Lista fotografija",
                "operationId": "d774d3ff2232e2859e57b0cf9ad69e73",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Fotografije"
                ],
                "summary": "Dodavanje fotografije (fotograf)",
                "operationId": "804eeb7eaf4974f11b05333ffa1a85df",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "naziv",
                                    "izlozba_id"
                                ],
                                "properties": {
                                    "naziv": {
                                        "type": "string",
                                        "example": "Planina"
                                    },
                                    "opis": {
                                        "type": "string",
                                        "example": "Opis..."
                                    },
                                    "izlozba_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "https://example.com/slika.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kreirano"
                    },
                    "403": {
                        "description": "Zabranjeno"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/fotografije/{id}": {
            "get": {
                "tags": [
                    "Fotografije"
                ],
                "summary": "Detalji fotografije",
                "operationId": "fef68180f89ba8cc47b7f259cfd1633a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Nije pronađena"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Fotografije"
                ],
                "summary": "Brisanje fotografije (admin)",
                "operationId": "42dbfebd0d1ec3dee0e0953859f91295",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obrisano"
                    },
                    "404": {
                        "description": "Nije pronađena"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Unesi token u formatu: Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}